requirements.txt
spacy~=3.4.0
gensim~=4.3.0
nltk~=3.6
tqdm~=4.64.0
PyYAML~=6.0
stanza==1.4.2
flask~=2.2.2
pandas==1.5.2
prance==0.22.11.4.0
openapi-spec-validator==0.5.1
mitmproxy==9.0.1
urllib3==1.26.7
requests~=2.30.0